Added exponential growth to fame

+++ tools/cython/KML/statistics/WindowedFame.pxd
@@ -6,6 +6,6 @@ cdef extern from "WindowedFame.cc":
 
 cdef extern from "WindowedFame.h" namespace "KML::Statistics":
     cdef cppclass WindowedFame:
-        WindowedFame(double, uint64_t) except +
+        WindowedFame(double, double) except +
         void update(double) except +
         double evaluate() except +
